Create joss_paper_draft.yml
Added the same GitHub action I wrote for XGA, in anticipation of setting up the paper structure for DAXA in the same way - a branch for writing a paper which can be merged into master when changes are made. Pushes to either paper/writeJOSSPaper or master will trigger the build action. 

Not that I have added a paper directory/files yet mind you.